:recycle: [Refactor] member 패키지 구조 변경에 따른 import 수정
